[Patchset 6] Add interactive trace_processor_shell
Adds an interactive shell (ninja target: trace_processor_shell)
to test the trace processor on the host.

+++ src/trace_processor/trace_processor_shell.cc
@@ -51,7 +51,7 @@ class FileReader : public BlobReader {
     return res > 0 ? static_cast<uint32_t>(res) : 0;
   }
 
-  size_t file_size() const { return file_size_; }
+  uint64_t file_size() const { return file_size_; }
 
  private:
   base::ScopedFile fd_;
